first

+++ src/main/java/com/sparta/hanghaememo/service/MemoService.java
@@ -1,7 +1,7 @@
 package com.sparta.hanghaememo.service;
 
+
 import com.sparta.hanghaememo.dto.MemoRequestDto;
-import com.sparta.hanghaememo.dto.MemoRequestDto2;
 import com.sparta.hanghaememo.entity.Memo;
 import com.sparta.hanghaememo.repository.MemoRepository;
 import lombok.RequiredArgsConstructor;
@@ -17,28 +17,41 @@ public class MemoService {
     private final MemoRepository memoRepository;
 
     @Transactional
-    public Memo createMemo(MemoRequestDto requestDto){ //Entity인 Memo객체를 반환
-        Memo memo = new Memo(requestDto); // Memo Entity 생성자에 클라이언트에서 넘겨준 데이터를 넣어줌
-        memoRepository.save(memo); //엔티티(테이블)을 리포지터리를 통해 저장
+    public Memo createMemo(MemoRequestDto requestDto) {
+
+        Memo memo = new Memo(requestDto);
+        memoRepository.save(memo);
         return memo;
     }
 
-    @Transactional(readOnly = true) //읽기 전용
+    @Transactional(readOnly = true)
     public List<Memo> getMemos() {
         return memoRepository.findAllByOrderByModifiedAtDesc();
     }
 
     @Transactional
-    public Long update(Long id, MemoRequestDto2 requestDto) {
-        Memo memo = memoRepository.findById(id).orElseThrow(  //업데이트할 메모가 있는지 확인
-                () -> new IllegalArgumentException("아이디가 존재하지 않습니다")
+    public Long update(Long id, MemoRequestDto requestDto) {
+        Memo memo = memoRepository.findById(id).orElseThrow(
+                () -> new IllegalArgumentException("아이디가 존재하지 않습니다.")
         );
-        memo.update(requestDto);
-        return memo.getId();
+        if (memo.getPassword().equals(requestDto.getPassword())) {
+            memo.update(requestDto);
+            return 123L;
+        }
+        else return -1L;
     }
+
     @Transactional
-    public Long deleteMemo(Long id) {
-        memoRepository.deleteById(id);
-        return id;
+    public Long deleteMemo(Long id, MemoRequestDto requestDto) {
+        Memo memo = memoRepository.findById(id).orElseThrow(
+                () -> new IllegalArgumentException("아이디가 존재하지 않습니다.")
+        );
+
+        if (memo.getPassword().equals(requestDto.getPassword())) {
+            memoRepository.deleteById(id);
+            return 1L;
+        }
+        else return -1L;
     }
+
 }
